% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpluce.R
\name{ordinal_seq}
\alias{ordinal_seq}
\title{Generate an Ordinal Sequence from a Number}
\usage{
ordinal_seq(n)
}
\arguments{
\item{n}{A single numeric value: the desired length of the sequence.}
}
\value{
A character vector of ordinal strings of length \code{n}.
}
\description{
Returns \code{c("1st","2nd","3rd", ...)} of length \code{n}.
}
\examples{
ordinal_seq(11)

}
