% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_dist.R
\name{print.rank_gap}
\alias{print.rank_gap}
\title{Print a rank_gap object}
\usage{
\method{print}{rank_gap}(x, ...)
}
\arguments{
\item{x}{A \code{rank_gap} object.}

\item{...}{Passed to methods.}
}
\description{
Print a rank_gap object
}
